Returning to standardized usage of Integer for Id fields

+++ backend/src/main/java/com/mealchemy/mealprep/dto/MealPlanEntryRequest.java
@@ -8,7 +8,7 @@ import java.time.*;
 import com.mealchemy.shared.enums.MealSlot;
 
 public record MealPlanEntryRequest(
-    @NotNull Long recipeId,
+    @NotNull Integer recipeId,
     @NotNull LocalDate entryDate,
     @NotNull MealSlot mealSlot,
     @NotNull LocalTime mealTime,
